"Add form validation and saving for make\_transaction view; remove stray backslash"
This commit adds form validation and saving functionality to the make\_transaction view. If the form is valid, it is saved and the user is redirected to the transaction\_list page. If the request method is not POST, an empty form is created and rendered in the make\_transactions template. A stray backslash at the end of the file has also been removed.