"Add Transaction model with associated fields and methods; import Account model"
This diff adds a new model called Transaction to the Transactions app, which includes fields for account, amount, type, transaction type, timestamp, and description. The Transaction model also includes a Meta class to specify the database table name and a \_\_str\_\_ method to provide a string representation of the object. Additionally, a static method get\_recent\_transactions is added to retrieve a specified number of recent transactions for a given account. The Account model from the Account app is also imported for use in the Transaction model.