"Add account detail view and retrieve associated transactions"
This commit adds a new view for account details and retrieves all transactions associated with the specified account number. The new view is called 'account\_detail' and takes 'account\_number' as a parameter. It retrieves the account object using the 'Account' model and filters all transactions associated with the account using the 'Transaction' model.