"Render account detail template with request context and add transactions"
This commit updates the `account_detail` view function in the `views.py` file of the `Account` app in the `Yomei-BankApp-master` project. It retrieves the account with the given account number and filters the transactions related to that account. Then, it returns the `account_detail.html` template with the `request` context and adds `transactions` as a template variable.