Added form and messages imports to views.py; no changes to TransactionForm or make\_transaction view; added messages context to deposit view.
The changes in this diff include adding the following imports to views.py:

* `from django.contrib import messages`
* `from .models import Transaction`
* `from .forms import TransactionForm`

There are no changes to the TransactionForm class or the make\_transaction view. However, the deposit view has been updated to include the `messages` context.

Note: The diff also includes the removal of duplicate imports at the beginning of views.py, but those imports are not relevant to the changes made in this diff.