"PyCU file changes in InfinityFinance views.pyc"
This commit message describes the code diff provided, which shows changes to a pycache file (views.cpython-310.pyc) in the InfinityFinance directory of the Yomei-BankApp project. PyCU files are byte-compiled files created by Python when a .py file is executed, so this diff shows changes to the compiled version of the views.py file. However, it's important to note that pyc files are not intended to be modified directly and are generally not human-readable, so further investigation would be needed to understand the specific changes made.